requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
pycryptodomex==3.14.1
pyserial==3.5
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0